feat(client): mirror VisAsset into standalone, satisfy drift parity (#93)
The standalone↔package drift test (tests/test_standalone_drift.py) parses
client.py via AST, so a class living in asset.py wasn't visible to the
parity check. Define VisAsset in client.py (alongside MtlxSource) and
have asset.py re-export it. User-facing imports are unchanged:

    from mat_vis_client import VisAsset                  # works
    from mat_vis_client.asset import VisAsset            # works
    from mat_vis_client.client import VisAsset           # works
    client.asset(src, mid, tier).to_threejs()            # works

Mirror VisAsset + MatVisClient.asset() into mat_vis_client_standalone.py
so the single-file client retains feature parity. test_standalone_drift's
4 tests pass.
