+++ clients/python/src/mat_vis_client/client.py
@@ -35,10 +35,6 @@ import urllib.request
 from importlib.metadata import PackageNotFoundError
 from importlib.metadata import version as _pkg_version
 from pathlib import Path
-from typing import TYPE_CHECKING
-
-if TYPE_CHECKING:
-    from mat_vis_client.asset import VisAsset
 
 REPO = "MorePET/mat-vis"
 GITHUB_API = f"https://api.github.com/repos/{REPO}"  # update-check only
@@ -1381,8 +1377,6 @@ class MatVisClient:
         Creation is free — no network IO until ``.scalars`` / ``.textures``
         / an adapter method is accessed. Mat-vis#93.
         """
-        from mat_vis_client.asset import VisAsset
-
         return VisAsset(self, source, material_id, tier)
 
     def _scalars_for(self, source: str, material_id: str) -> dict:
@@ -1886,6 +1880,155 @@ class MtlxSource:
             self._tier,
             is_original=True,
         )
+
+
+# ── VisAsset ────────────────────────────────────────────────────
+
+_VIS_ASSET_FROZEN = ("_source", "_material_id", "_tier")
+
+
+class VisAsset:
+    """Bundle of (identity + lazy scalars + lazy textures + adapters).
+
+    Two-layer shape (mat-vis#93): VisAsset is the ergonomic class; the
+    underlying ``to_threejs`` / ``to_gltf`` / ``export_mtlx`` free functions
+    in :mod:`mat_vis_client.adapters` remain the stable primitives. Adapter
+    methods on this class call those primitives with identity-bound args.
+
+    Mirrors ``requests.Session`` / ``requests.get()`` and
+    ``subprocess.Popen`` / ``subprocess.run()``: class as ergonomic
+    surface, free functions as port-friendly primitives.
+
+    Identity is **immutable** — assigning to ``source``/``material_id``/``tier``
+    raises ``AttributeError``. Use :meth:`with_tier` to spawn a new
+    instance for a different tier. Equality and hashing are identity-only
+    (same triple, even across distinct client instances).
+    """
+
+    __slots__ = (
+        "_client",
+        "_source",
+        "_material_id",
+        "_tier",
+        "_scalars_cache",
+        "_textures_cache",
+        "_initialized",
+    )
+
+    def __init__(
+        self,
+        client: MatVisClient,
+        source: str,
+        material_id: str,
+        tier: str,
+    ) -> None:
+        object.__setattr__(self, "_client", client)
+        object.__setattr__(self, "_source", source)
+        object.__setattr__(self, "_material_id", material_id)
+        object.__setattr__(self, "_tier", tier)
+        object.__setattr__(self, "_scalars_cache", None)
+        object.__setattr__(self, "_textures_cache", None)
+        object.__setattr__(self, "_initialized", True)
+
+    @classmethod
+    def from_client(
+        cls,
+        client: MatVisClient,
+        source: str,
+        material_id: str,
+        tier: str = "1k",
+    ) -> "VisAsset":
+        return cls(client, source, material_id, tier)
+
+    def __setattr__(self, name: str, value: object) -> None:
+        if getattr(self, "_initialized", False) and name in _VIS_ASSET_FROZEN:
+            raise AttributeError(f"{name} is immutable; use with_tier() / new VisAsset")
+        object.__setattr__(self, name, value)
+
+    @property
+    def source(self) -> str:
+        return self._source
+
+    @property
+    def material_id(self) -> str:
+        return self._material_id
+
+    @property
+    def tier(self) -> str:
+        return self._tier
+
+    def with_tier(self, tier: str) -> "VisAsset":
+        """Return a new :class:`VisAsset` with the same source/material_id but a different tier."""
+        return VisAsset(self._client, self._source, self._material_id, tier)
+
+    @property
+    def scalars(self) -> dict:
+        """Lazy PBR scalars (cached after first access).
+
+        Calls :meth:`MatVisClient._scalars_for` exactly once per instance.
+        """
+        if self._scalars_cache is None:
+            object.__setattr__(
+                self,
+                "_scalars_cache",
+                self._client._scalars_for(self._source, self._material_id),
+            )
+        return self._scalars_cache
+
+    @property
+    def textures(self) -> dict[str, bytes]:
+        """Lazy channel -> PNG bytes mapping (cached after first access).
+
+        Calls :meth:`MatVisClient.fetch_all_textures` exactly once per instance.
+        """
+        if self._textures_cache is None:
+            object.__setattr__(
+                self,
+                "_textures_cache",
+                self._client.fetch_all_textures(self._source, self._material_id, self._tier),
+            )
+        return self._textures_cache
+
+    def to_threejs(self) -> dict:
+        """Return a Three.js ``MeshPhysicalMaterial`` parameter dict.
+
+        Wraps :func:`mat_vis_client.adapters.to_threejs` with this asset's
+        identity-bound scalars and textures.
+        """
+        from mat_vis_client.adapters import to_threejs
+
+        return to_threejs(self.scalars, self.textures)
+
+    def to_gltf(self) -> dict:
+        """Return a glTF 2.0 material dict.
+
+        Wraps :func:`mat_vis_client.adapters.to_gltf` with this asset's
+        identity-bound scalars and textures.
+        """
+        from mat_vis_client.adapters import to_gltf
+
+        return to_gltf(self.scalars, self.textures)
+
+    def to_mtlx(self) -> "MtlxSource":
+        """Return a fresh :class:`MtlxSource` for this asset's identity.
+
+        Composition, not replacement: ``MtlxSource`` remains the public
+        MaterialX façade. This is the recommended entry point.
+        """
+        return MtlxSource(self._client, self._source, self._material_id, self._tier)
+
+    def __eq__(self, other: object) -> bool:
+        return isinstance(other, VisAsset) and (
+            self._source,
+            self._material_id,
+            self._tier,
+        ) == (other._source, other._material_id, other._tier)
+
+    def __hash__(self) -> int:
+        return hash((self._source, self._material_id, self._tier))
+
+    def __repr__(self) -> str:
+        return f"VisAsset({self._source!r}, {self._material_id!r}, tier={self._tier!r})"
 
 
 def _render_synthesized_mtlx_xml(
